// src/adapter/generated-types.ts
export interface Miner {
  version: number;
  bump: number;
  authority: PublicKey;
  unclaimed_usd_amount: BN;
  unclaimed_btc_shares: BN;
  hashrate_amount: BN;
  current_streak_count: number;
  last_mined_round_id: number;
  unclaimed_hashrate: BN;
  grubstake_usd_amount: BN;
  grubstake_expiration_timestamp: BN;
  affiliate: PublicKey;
  unclaimed_token_shares: BN;
  reserved: number[];
}
// ...
export interface SatsVault {
  version: number;
  bump: number;
  btc_amount: BN;
  btc_shares: BN;
  leftovers: BN;
  reserved: number[];
}

export interface TileStake {
  stake: BN;
  deploy_count: number;
}

export interface TokenClaimed {
  authority: PublicKey;
  claimed_shares: BN;
  tokens_received: BN;
  claimed_sats_shares: BN;
  btc_received: BN;
  claimed_hashrate: BN;
}

export interface TokenVault {
  version: number;
  bump: number;
  token_amount: BN;
  token_shares: BN;
  leftovers: BN;
  reserved: number[];
}